Add more HRESULT checks
On multiple places, there is an interaction with endpointVolume and the function is expected to return something, 
but the return value of endpointVolume is not checked. Therefore, undefined values may be returned.
This commit adds return value checks with default values as return values for failing endpointVolume interactions.
This should fix null pointer dereference issues.